Enemies are now attached to packs the moment a pack is drawn over them. Should there be two packs, it'll just take over the enemy. Though packs should never overlap.
AdminEnemy had some issues with if .enemypack.id should be used or .enemy_pack_id, latter won this battle. We don't know the full pack when loading in fresh; nor do we need to
AdminEnemyPack was missing setSynced calls
enemyattaching (still hate the name but w/e) now registers for whenever enemy packs are created, when so, it'll register to its synced event (we now know its ID), then it'll loop over enemies to see who is caught in it, if so, that enemy is then attached to the pack. Whew.
DungeonMap: clicking gets rid of the annoying 'ctrl to scroll blabla' message
MapObject: now triggers a synced event when synced was true. Perhaps need to trigger it on false as well, but no need for now.
Signalable: added dummy unregister function. No real reliable way of removing just one reference, but I want to get the function in and call it so I can implement it some other day.