Added WIP DungeonDataSeeder which imports from exported .json files per dungeon/floor. Added Blizzard copyright to tiles.
So the idea is that I edit enemies/packs/younameit on my dev instance, hit an export button, and out comes rolling all these files with exports per dungeon floor which are then checked in GIT. On production, these files will be fetched and a seeder is ran to clear the instance of enemies/packs, then re-fill the entire database as it was on my dev server. This way there's a nice changelog in GIT about what enemies etc there are and how they're updated as the website evolves. It's also easier to see where these updates are occurring and it's easy to revert to a version that worked if I ever bork something.