#59 Fixed and added cloning ability in full. Closes #59.
Added a special column to keep track of which routes have been cloned by which routes. This is also displayed if the route you cloned it from is not your own to automatically give credit where credit is due. Added clone button to profile dungeon route list to make it easier to clone existing routes of your own.

Removed ads from the home page for the time being.