#39 Added infested supported for modifiers with accompanying adjustments to the EnemyVisual system. Added a new 'infested voting' (name subject to change) layer which allows a user to vote all existing mobs for Infesting or not (not functional atm).
Use interface is almost done, perhaps some touch ups still, but other than that I can work on tying the front end and the back-end together. I'll still need a way to visualize how many votes are required for turning an enemy into an Infested enemy. And also work on the back end a bit so that Infested votes of a previous week are disregarded (not discarded yet!) AND need to add a Region toggle so that people in the EU will get served Infested enemies a day longer than their US counterparts.