#40 Fixed a multitude of things related to splitting up the edit and the view sidebars.
Virtual tour works again with the new layout.
Created some new global functions.
Every time you change the enemy forces, a small popup comes and notifies you of it.
Fixed an issue where Infested Voting information was not available on the tryout version. Voting is still not possible if you're not logged in.
Fixed an issue where you could attempt to assign raid markers to enemies in view mode (didn't work, but still). I also think I re-introduced this bug so hopefully it stays gone for a while now.