Added MDT export string decoding to a PHP array and vice versa, closes #96
I initially tried to convert the LUA implementation to PHP but found out that that wasn't going to work. The Huffman encoding library didn't seem to produce the output I was expecting (or perhaps the decoding), then even if it did I'd ran into the issue of the deserializing. In the end I probably could have done it, but going the LUA route also made sure the code wasn't going to be maintained by me, which always helps.

I added some hints and pointers to docs/lua.md for when I ever decide to reinstall the current server where Keystone.guru is running on (bless me when that happens) so I can get LUA working there.

Kept the PHP implementation of what I have for the time being, perhaps it'll come in useful later on.