#110 A couple of high-usage handlebars have been precompiled now, this saves a LOT of JS loading time. The page is visibly faster now which is great.
Not all templates have been converted, there's a bunch of PHP in some templates which are hardly used. For now I don't think it's wise to convert those purely for the sake of converting, the performance gains from those will be minimal. A lot of other small non-PHP templates have been converted.