#159 Added diff tool for KG and MDT and processed all differences that the code found. This fixes quite some bugs granted that MDT is correct.
The diff tool is accessed from a new page in the admin called "Admin tools". In here, some specific admin related jobs can be found, including a tool for dumping an MDT string for dissecting and exporting enemy data from database to file.

I also took the liberty to add a adminpanel control to AdminDungeonMap which for now only shows the coordinates of your current mouse pointer position. This helps with debugging enemy locations.