#77 Added team page and cleaned it up a little.
Teams page now is tabbed and contains ways to edit the team and show the invite link/show current members.
From that page, one should be able to promote/demote users though that isn't implemented yet.
Invite code accept page is largely done, probably some touch ups needed still but good enough as it is.
Actually getting into a group is not functional at this time, nor is adding routes to a team.