#37 Various fixes that popped up when putting v2.5 live.
Live and staging were meddling in the same REDIS database, this has been fixed.
You were not able to add private routes to a team, this has been fixed.
Losing connection to the server while cooperatively editing a route would cause the user list to not be cleared and be duplicated. This has been fixed.
Added a missing changelog.