#179 Added creative-tim Argon dashboard.
The dashboard shows details related to users, teams, routes and pageviews. It gives me an overview of how the website is doing.
I've renamed the EchoServerServiceProvider to KeystoneGuruServiceProvider; this is a single service provider for the entire website and not just for Echo Server related things.
Added a bunch of services for the Dashboard, mainly related to statistics for the dashboard.