#185 Opted not to add any admin pages.
Instead, I've reworked the back-end to use Seasons rather than something injected in the GameServerRegion.
There's a new Service around which you ask when you want anything to do with M+ Seasons. Then, you can use the Season to resolve specifics for the information gathered in the Service.
Fixed an issue where filtering the routes table was not working anymore.