#184 Added Beguiling support
Beguiling enemies are added to an EnemyPack, where enemies are then saved into the database as if a normal enemy.
They're saved like this so all existing functionality works for them too (such as raid markers)
Based on the currently displayed preset, enemies are displayed or hidden from the user.