#213 Improved enemy class display.
There was some minor pixel fuckery going on which I solved by using the calc() function.
Upscaled the boss icon so it's no longer a pixelated mess.
Reduced max size and margin to have less pixels wasted on screen. The point comes across as-is; enemies with more hp are bigger on the map.