#344 Various issues with the MDT import fixed.
I have fixed a few issues with this string. I removed the warning where it says 'you configured stuff for A and B weeks, I'm choosing whatever comes first' since the MDT string simply defines for which week the route is, so I can choose the correct one always (not an issue with this string per-se). When importing a string the A/B pillar bosses would not show up until a refresh, this is fixed. I need to check in-game how this string looks but on my side it looks like this string has the selected affix of (say) an A week, but only pulls with B week pillar bosses, thus not importing the pillar bosses correctly. I'd need to fire up MDT to see how it looks there, may be an issue on my side or the string.

Pulls that end up empty for whatever reason (pull for B week pillar bosses but string says use A week) are now properly deleted rather than getting an empty pull.

Finally I'll need to build some things to properly draw the lines from the pillars. If pillars switch due to A/B week it may draw a line from the wrong pillar.