#430 It's okay to use multiple artisan commands in the pipeline.
This is cleaner for when running it for local, pipeline will just perform multiple commands. Not a problem in itself. Pipeline will now pull the code first so we have new changes, and THEN execute the updated commands. Prevents the environment from getting stuck in bugs that may have been pushed before.