#1251 Fixed various issues with creating routes for a season, rather than an expansion.
An expansion contains a list of dungeons, and up until now the list of dungeons for a season was tied to whatever was there in an expansion. With SL S4 (and DF S1) this is no longer the case. A season can consist of dungeons from any expansion.
I had already created support for this, but now the site needs to show what dungeons are currently available in the season and the next season (if applicable).