#1335 Mapping versions are now saved and loaded properly. All entities that are versioned are now also no longer dropped from the database.
I still need to expand on MappingVersionConditional.php where it'd check if this model is a newer version, and if so allow it to be inserted in the database. For now it denies everything.
Lastly, I moved a lot of the relation import stuff into the App folder so that it'd be properly PSR recognized.