#2447 Reworked how spells are extracted from a combat log and output to csv.
A combat log or MDT is a way to extract spells from the game. If a new spell is found, it is saved to the database. Any new info is extracted from Wowhead.

Npcs cast spells, and spells are assigned to 1 or more dungeons (that NPCs live in).

The csv extractor simply looks at what dungeon you want the spells for, and outputs it to file. It no longer requires a combat log. As a result, I can now also "combine" spell sources and lump them together to build a more comprehensive .csv file.