#2464 Updated Seeders to choose when to run which seeders based on environment.
This prevents a clash when staging+testing are deployed at the same time, but they're seeding the same tables since the combatlog database is shared across environments (maybe not the best idea in the end but it is what it is).

Added better support for Evokers and added the Earthen as a race.