#2676 ui_map_id is no longer validated with the database when sending ARC requests or corrections.
Blizzard does not send consistent IDs in that field so to prevent stuff from not working I just ensure that floors are resolved otherwise. ui_map_id still takes precedence, but otherwise the previously known floor is used, and if that is not known we use the default floor for the dungeon.

I added additional tests to ensure that POST bodies with errors still get parsed correctly.