#2831 Thumbnail generating reworked to now upload into S3 instead.
This is required for the server migration. Existing thumbnails will be WIPED. This is on purpose - it generated way too many thumbnails (600k!) totaling 30GB worth of data.

I changed the logic to only generate thumbnails for those routes that actually are viewed in various list views, and those routes that are popular (ie. have views in the first place).