#2945 Disabled model cache when exporting MDT strings.
There was an odd issue where pulls and notes from one route was exported to another route. Disabling the model cache should fix this.

I've also added a local cache of strings for exporting. This is the replacement cache which should work a lot better and reduce load drastically.

When in edit mode you always bypass the cache just to be sure.